"Refactor grpc services with client-server interaction improvement."
This commit contains a large-scale refactor of the existing grpc services codebase with significant improvements associated to the client-server interaction. Numerous files have been deleted, added, or modified. Key changes include replacing "DisconnectInterceptor" with "ConnectionInterceptor", introducing a link between the client and server for heartbeats with the addition of "HeartBeatTask" and "HeartBeatService", and creating new gRPC services "HelloService". Also, a new proto file ("exchange.proto") for exchange services has been added. This commit aims to improve the performance, consistency, and readability of the code.