Refactor GRPC service and enhance monitoring
This commit primarily handles the renaming and refactoring of several classes related to the GRPC server and client configuration. A '@Slf4j' annotation was added to enable logging. In the newly created GrpcServerTransportFilterConfig, a MonitoringServerTransportFilter is setup to monitor server transport events. Meanwhile, the protobuf file "exchange.proto" was modified to change the 'report' RPC method signature. At the client-side, scheduled heartbeat reporting was implemented to the server.