Refactor HeartBeat service and task to use Report proto
The original Exchange-related code for heartbeat service and task has been refactored to use Report proto instead. This includes modifying dependencies in HeartBeatService and HeartBeatTask, and creating a new ReportService in the proto file while removing the heartbeat services from the ExchangeService. Also, classes have been renamed in accordance with the new Report service and proto.