Update application names and modify gRPC endpoint response
The application names in both client and server 'application.yml' files have been updated to more appropriately reflect their roles. In addition, the response of the '/' endpoint in 'HttpTestController.java' has been changed from a simple string to a map for more diverse output. Now it returns not just the server name, but also its readiness and connection status.